change v2/filters to v2/filterMaps; rename routines in UrlOverview
