+++ src/components/shared/filters.css
@@ -68,7 +68,7 @@
                 /*.filter-box {*/
                 /*    !*margin-left:1rem;*!*/
                 /*}*/
-                /*.filter-box.ref-filters-0 {*/
+                /*.filter-box.ref-filter-0 {*/
                 /*    margin-left:0;*/
                 /*}*/
                 /*.filter-box h4 {*/
